Add basic postgres integration
