+++ ftl/ftl-module-offerings/src/main/kotlin/ftl/offerings/Offerings.kt
@@ -1,17 +1,19 @@
 package ftl.offerings
 
-import ftl.lib.Json
+import ftl.lib.PostgresClient
 import xyz.block.ftl.Context
-import xyz.block.ftl.Ingress
-import xyz.block.ftl.Method
 import xyz.block.ftl.Verb
 
 data class OfferingsRequest(val name: String)
 data class OfferingsResponse(val message: String)
 
 class Offerings {
+  private var postgresClient = PostgresClient()
+
   @Verb
   fun echo(context: Context, req: OfferingsRequest): OfferingsResponse {
-    return OfferingsResponse(message = "Hello, ${req.name}! ${Json.stringify(req)}")
+    postgresClient.ping()
+    postgresClient.close()
+    return OfferingsResponse(message = "Hello, ${req.name}!")
   }
 }
